% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/BlessyM.R, R/Power.R
\name{Blessy}
\alias{Blessy}
\title{Square a number}
\usage{
Blessy(x)

Blessy(x)
}
\arguments{
\item{x}{A numeric value to be squared}

\item{x}{A numeric value to be squared}
}
\value{
The square of teh input

The square of teh input
}
\description{
Takes in any numberis value and square it .

Takes in any numberis value and square it .
}

