% Generated by roxygen2 (4.1.0): do not edit by hand
<<<<<<< HEAD
% Please edit documentation in R/BlessyM.R
=======
% Please edit documentation in R/Power.R
>>>>>>> d88941f0e75a5faf89717c77fa700ca978042c80
\name{Blessy}
\alias{Blessy}
\title{Square a number}
\usage{
Blessy(x)
}
\arguments{
\item{x}{A numeric value to be squared}
}
\value{
The square of teh input
}
\description{
Takes in any numberis value and square it .
}

